tests/wpt/web-platform-tests/workers/constructors/SharedWorker/name.html

<!--
onconnect = function(e) {
  e.ports[0].postMessage(self.name);
}
/*
-->
<!doctype html>
<title>self.name</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
(async_test()).step(function() {
  var worker = new SharedWorker('#', 'hello');
  worker.port.addEventListener('message', this.step_func(function(e) {
    assert_equals(e.data, 'hello');
    this.done();
  }), false);
  worker.port.start();
});
</script>
<!--
*/
//-->
